+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/BuildDockerTask.java
@@ -103,14 +103,6 @@ public class BuildDockerTask extends DefaultTask implements JibTask {
     Preconditions.checkNotNull(jibExtension);
     TaskCommon.disableHttpLogging();
 
-    if (jibExtension.getUseOnlyProjectCache()) {
-      getLogger()
-          .warn(
-              "jib.useOnlyProjectCache is deprecated; use the 'jib.useOnlyProjectCache' system "
-                  + "property instead, or set the cache directories using the 'jib.baseImageCache' "
-                  + "and 'jib.applicationCache' system properties.");
-    }
-
     try {
       AbsoluteUnixPath appRoot = TaskCommon.getAppRootChecked(jibExtension, getProject());