+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/BuildImageTask.java
@@ -80,14 +80,6 @@ public class BuildImageTask extends DefaultTask implements JibTask {
     Preconditions.checkNotNull(jibExtension);
     TaskCommon.disableHttpLogging();
 
-    if (jibExtension.getUseOnlyProjectCache()) {
-      getLogger()
-          .warn(
-              "jib.useOnlyProjectCache is deprecated; use the 'jib.useOnlyProjectCache' system "
-                  + "property instead, or set the cache directories using the 'jib.baseImageCache' "
-                  + "and 'jib.applicationCache' system properties.");
-    }
-
     try {
       AbsoluteUnixPath appRoot = TaskCommon.getAppRootChecked(jibExtension, getProject());
       GradleProjectProperties projectProperties =