+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/BuildDockerMojo.java
@@ -61,14 +61,6 @@ public class BuildDockerMojo extends JibPluginConfiguration {
       return;
     }
 
-    if (getUseOnlyProjectCache()) {
-      getLog()
-          .warn(
-              "<useOnlyProjectCache> is deprecated; use the 'jib.useOnlyProjectCache' system "
-                  + "property instead, or set the cache directories using the 'jib.baseImageCache' "
-                  + "and 'jib.applicationCache' system properties.");
-    }
-
     if (!DockerClient.isDefaultDockerInstalled()) {
       throw new MojoExecutionException(
           HelpfulSuggestions.forDockerNotInstalled(HELPFUL_SUGGESTIONS_PREFIX));