+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/BuildImageMojo.java
@@ -64,14 +64,6 @@ public class BuildImageMojo extends JibPluginConfiguration {
       return;
     }
 
-    if (getUseOnlyProjectCache()) {
-      getLog()
-          .warn(
-              "<useOnlyProjectCache> is deprecated; use the 'jib.useOnlyProjectCache'  system "
-                  + "property instead, or set the cache directories using the 'jib.baseImageCache' "
-                  + "and 'jib.applicationCache' system properties.");
-    }
-
     // Validates 'format'.
     if (Arrays.stream(ImageFormat.values()).noneMatch(value -> value.name().equals(getFormat()))) {
       throw new MojoFailureException(