+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/BuildTarMojo.java
@@ -61,14 +61,6 @@ public class BuildTarMojo extends JibPluginConfiguration {
       return;
     }
 
-    if (getUseOnlyProjectCache()) {
-      getLog()
-          .warn(
-              "<useOnlyProjectCache> is deprecated; use the 'jib.useOnlyProjectCache' system "
-                  + "property instead, or set the cache directories using the 'jib.baseImageCache' "
-                  + "and 'jib.applicationCache' system properties.");
-    }
-
     try {
       AbsoluteUnixPath appRoot = MojoCommon.getAppRootChecked(this);
       MavenProjectProperties projectProperties =