+++ jib-plugins-common/src/main/java/com/google/cloud/tools/jib/plugins/common/RawConfiguration.java
@@ -67,7 +67,5 @@ public interface RawConfiguration {
 
   boolean getAllowInsecureRegistries();
 
-  boolean getUseOnlyProjectCache();
-
   ImageFormat getImageFormat();
 }