+++ jib-maven-plugin/CHANGELOG.md
@@ -11,7 +11,7 @@ All notable changes to this project will be documented in this file.
 
 - `gwt-app` packaging type now builds a WAR container.
 - When building to Docker and no `<to><image>` is defined, artifact ID is used as an image reference instead of project name.
-- `<useOnlyProjectCache>` parameter is deprecated in favor of the `jib.useOnlyProjectCache` system property ([#1308](https://github.com/GoogleContainerTools/jib/issues/1308))
+- Removed `<useOnlyProjectCache>` parameter in favor of the `jib.useOnlyProjectCache` system property ([#1308](https://github.com/GoogleContainerTools/jib/issues/1308))
 
 ### Fixed